style: review comments
