+++ pkg/process/process.go
@@ -18,7 +18,9 @@ const (
 // - tanka.dev/** labels
 // - filtering
 // - best-effort sorting
-func Process(raw interface{}, cfg v1alpha1.Environment, exprs Matchers) (manifest.List, error) {
+func Process(cfg v1alpha1.Environment, exprs Matchers) (manifest.List, error) {
+	raw := cfg.Data
+
 	if raw == nil {
 		return manifest.List{}, nil
 	}
